requirements.txt
aiohttp==3.8.6
pytest==7.4.4
pyyaml==6.0.1
requests==2.31.0
urllib3==1.26.7
